Refactor QuantBatchNorm2d into QuantScaleBias and BatchNorm2dToQuantScaleBias
