+++ brevitas/nn/__init__.py
@@ -3,4 +3,5 @@ from .quant_activation import QuantReLU, QuantSigmoid, QuantTanh, QuantHardTanh
 from .quant_avg_pool import QuantAvgPool2d
 from .quant_linear import QuantLinear
 from .quant_conv import QuantConv2d, PaddingType
-from .quant_bn import QuantBatchNorm2d
+from .quant_bn import BatchNorm2dToQuantScaleBias
+from .quant_scale_bias import ScaleBias, QuantScaleBias
